Proper modem startup on ZR
Since the ZR is a APQ and not a MSM device, we have to properly
initialize its modem to get radio connectivity. This means that we need
a couple of blobs in place. For those to work, we need proper
permissions on devices files. We also fix the bogus path to the
kickstart script.